feat: Move httprequest mocking, finish unit tests
